+++ broker/src/functional-test/groovy/com/swisscom/cloud/sb/broker/functional/BindingParametersFunctionalSpec.groovy
@@ -3,6 +3,7 @@ package com.swisscom.cloud.sb.broker.functional
 import com.swisscom.cloud.sb.broker.model.repository.ServiceBindingRepository
 import com.swisscom.cloud.sb.broker.services.common.ServiceProviderLookup
 import com.swisscom.cloud.sb.broker.util.test.DummySynchronousServiceProvider
+import com.swisscom.cloud.sb.client.model.DeleteServiceInstanceBindingRequest
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.HttpStatus
 import org.springframework.web.client.HttpClientErrorException
@@ -78,6 +79,41 @@ class BindingParametersFunctionalSpec extends BaseFunctionalSpec {
         bindingResponse.body.parameters != null
     }
 
+    def "provision async service instance and fetch non existing binding"() {
+        given:
+        serviceLifeCycler.createServiceIfDoesNotExist('SyncDummyInstancesRetrievable', ServiceProviderLookup.findInternalName(DummySynchronousServiceProvider.class), null, null, null, 0, true, true)
+
+        def serviceInstanceGuid = UUID.randomUUID().toString()
+        def serviceBindingGuid = UUID.randomUUID().toString()
+        serviceLifeCycler.setServiceInstanceId(serviceInstanceGuid)
+        serviceLifeCycler.setServiceBindingId(serviceBindingGuid)
+        serviceLifeCycler.createServiceInstanceAndAssert(0, false, false,)
+
+        when:
+        serviceBrokerClient.getServiceInstanceBinding(serviceInstanceGuid, serviceBindingGuid)
+
+        then:
+        def ex = thrown(HttpClientErrorException)
+        ex.statusCode == HttpStatus.NOT_FOUND
+    }
+
+    def "provision async service instance and unbind non existing binding"() {
+        given:
+        serviceLifeCycler.createServiceIfDoesNotExist('SyncDummyInstancesRetrievable', ServiceProviderLookup.findInternalName(DummySynchronousServiceProvider.class), null, null, null, 0, true, true)
+
+        def serviceInstanceGuid = UUID.randomUUID().toString()
+        def serviceBindingGuid = UUID.randomUUID().toString()
+        serviceLifeCycler.setServiceInstanceId(serviceInstanceGuid)
+        serviceLifeCycler.setServiceBindingId(serviceBindingGuid)
+        serviceLifeCycler.createServiceInstanceAndAssert(0, false, false,)
+
+        when:
+        serviceBrokerClient.deleteServiceInstanceBinding(new DeleteServiceInstanceBindingRequest(serviceInstanceGuid,
+                serviceBindingGuid, serviceLifeCycler.cfService.guid, serviceLifeCycler.cfService.plans[0].guid))
+        then:
+        def ex = thrown(HttpClientErrorException)
+        ex.statusCode == HttpStatus.GONE
+    }
 
     def "deprovision async service instance"() {
         when: