+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/ServiceBroker.groovy
@@ -10,18 +10,18 @@ import org.springframework.boot.builder.SpringApplicationBuilder
 import org.springframework.boot.web.support.SpringBootServletInitializer
 
 @SpringBootApplication(exclude = [org.springframework.cloud.servicebroker.config.ServiceBrokerAutoConfiguration.class,
-        MongoAutoConfiguration.class, MongoDataAutoConfiguration.class])
+		org.springframework.boot.autoconfigure.mongo.MongoAutoConfiguration, org.springframework.boot.autoconfigure.data.mongo.MongoDataAutoConfiguration])
 @CompileStatic
 class ServiceBroker extends SpringBootServletInitializer {
-	static void main(String[] args) {
-		new SpringApplicationBuilder(ServiceBroker.class)
-				.initializers(new StandAloneConfigurationInitializer()).run(args)
-	}
+    static void main(String[] args) {
+        new SpringApplicationBuilder(ServiceBroker.class)
+                .initializers(new StandAloneConfigurationInitializer()).run(args)
+    }
 
 	@Override
 	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
 		logger.info("Initializing service broker")
 		//Needed for war based deployment
-		application.initializers(new WebContainerConfigurationInitializer()).sources(ServiceBroker)
+		application.initializers(new WebContainerConfigurationInitializer()).sources(ServiceBroker).profiles("influx")
 	}
 }