+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/controller/BindingController.groovy
@@ -1,16 +1,13 @@
 package com.swisscom.cloud.sb.broker.controller
 
 import com.fasterxml.jackson.databind.ObjectMapper
-import com.swisscom.cloud.sb.broker.binding.BindRequest
-import com.swisscom.cloud.sb.broker.binding.BindResponse
-import com.swisscom.cloud.sb.broker.binding.FetchServiceBindingProvider
-import com.swisscom.cloud.sb.broker.binding.ServiceBindingPersistenceService
-import com.swisscom.cloud.sb.broker.binding.ServiceInstanceBindingResponseDto
-import com.swisscom.cloud.sb.broker.binding.UnbindRequest
+import com.swisscom.cloud.sb.broker.binding.*
 import com.swisscom.cloud.sb.broker.cfapi.converter.ServiceInstanceBindingDtoConverter
 import com.swisscom.cloud.sb.broker.cfapi.dto.BindRequestDto
 import com.swisscom.cloud.sb.broker.cfapi.dto.UnbindingDto
 import com.swisscom.cloud.sb.broker.error.ErrorCode
+import com.swisscom.cloud.sb.broker.metrics.BindingMetricsService
+import com.swisscom.cloud.sb.broker.error.ServiceBrokerException
 import com.swisscom.cloud.sb.broker.model.CFService
 import com.swisscom.cloud.sb.broker.model.Plan
 import com.swisscom.cloud.sb.broker.model.ServiceBinding
@@ -26,11 +23,7 @@ import io.swagger.annotations.ApiOperation
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RequestMethod
-import org.springframework.web.bind.annotation.RestController
+import org.springframework.web.bind.annotation.*
 
 import javax.validation.Valid
 import java.security.Principal
@@ -51,6 +44,9 @@ class BindingController extends BaseController {
     @Autowired
     private ServiceInstanceBindingDtoConverter bindingDtoConverter
 
+    @Autowired
+    private BindingMetricsService bindingMetricsService
+
     @ApiOperation(value = "Bind service")
     @RequestMapping(value = '/v2/service_instances/{service_instance}/service_bindings/{id}', method = RequestMethod.PUT)
     ResponseEntity<String> bind(@PathVariable('id') String bindingId,
@@ -60,6 +56,7 @@ class BindingController extends BaseController {
         log.info("Bind request for bindingId: ${bindingId}, serviceId: ${bindingDto?.service_id} and serviceInstanceGuid: ${serviceInstanceId}")
 
         ServiceInstance serviceInstance = getAndCheckServiceInstance(serviceInstanceId)
+        bindingMetricsService.setTotalBindingRequestsPerService(serviceInstance)
         verifyServiceInstanceIsReady(serviceInstance) //Don't allow binding if service is not ready
         CFService service = getAndCheckService(bindingDto)
         failIfServiceBindingAlreadyExists(bindingId)
@@ -68,6 +65,7 @@ class BindingController extends BaseController {
 
         serviceBindingPersistenceService.create(serviceInstance, getCredentialsAsJson(bindResponse), serializeJson(bindingDto.parameters), bindingId, bindResponse.details, bindingDto.context, principal.name)
 
+        bindingMetricsService.setSuccessfulBindingRequestsPerService(serviceInstance)
         return new ResponseEntity<String>(getCredentialsAsJson(bindResponse), bindResponse.isUniqueCredentials ? HttpStatus.CREATED : HttpStatus.OK)
     }
 
@@ -92,7 +90,10 @@ class BindingController extends BaseController {
         if (!service) {
             ErrorCode.SERVICE_NOT_FOUND.throwNew("no service with id:${bindingDto.service_id} found")
         }
-        //TODO check if service is bindable
+
+        if (!service.bindable) {
+            ErrorCode.SERVICE_NOT_BINDABLE.throwNew("the service with the id:${bindingDto.service_id} is not bindable")
+        }
         return service
     }
 
@@ -120,7 +121,13 @@ class BindingController extends BaseController {
                @PathVariable('id') String bindingGuid,
                UnbindingDto unbindingDto) {
         log.info("Unbind request for BindingId: ${bindingGuid} and ServiceInstanceid: ${serviceInstanceId}")
-        ServiceBinding serviceBinding = checkServiceBinding(bindingGuid)
+        ServiceBinding serviceBinding
+        try {
+            serviceBinding = checkServiceBinding(bindingGuid)
+        } catch (ServiceBrokerException e) {
+            if (e.code == ErrorCode.SERVICE_BINDING_NOT_FOUND.code) ErrorCode.SERVICE_BINDING_GONE.throwNew()
+            else throw e
+        }
         ServiceInstance serviceInstance = serviceBinding.serviceInstance
         CFService service = serviceInstance.plan.service
 
@@ -154,7 +161,7 @@ class BindingController extends BaseController {
     private ServiceBinding checkServiceBinding(String bindingGuid) {
         ServiceBinding serviceBinding = serviceBindingRepository.findByGuid(bindingGuid)
         if (!serviceBinding) {
-            ErrorCode.SERVICE_BINDING_GONE.throwNew()
+            ErrorCode.SERVICE_BINDING_NOT_FOUND.throwNew()
         }
         return serviceBinding
     }