+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/services/kubernetes/templates/constants/AbstractTemplateConstants.groovy
@@ -4,7 +4,7 @@ import groovy.transform.CompileStatic
 
 @CompileStatic
 trait AbstractTemplateConstants {
-    private final String value
+    String value
 
     String getValue() {
         return this.value