+++ broker/src/main/resources/db/migration/V1_0_20__add_dateDeleted_ServiceInstance.sql
@@ -1,0 +1,2 @@
+ALTER TABLE service_instance
+  ADD date_deleted DATETIME;