+++ client/src/main/groovy/com/swisscom/cloud/sb/client/IServiceBrokerClientExtended.groovy
@@ -3,6 +3,7 @@ package com.swisscom.cloud.sb.client
 import com.swisscom.cloud.sb.model.backup.BackupDto
 import com.swisscom.cloud.sb.model.backup.RestoreDto
 import com.swisscom.cloud.sb.model.endpoint.Endpoint
+import com.swisscom.cloud.sb.model.health.ServiceHealth
 import com.swisscom.cloud.sb.model.usage.ServiceUsage
 import groovy.transform.CompileStatic
 import org.springframework.http.ResponseEntity
@@ -11,6 +12,8 @@ import org.springframework.http.ResponseEntity
 interface IServiceBrokerClientExtended extends IServiceBrokerClient {
     ResponseEntity<Endpoint> getEndpoint(String serviceInstanceId)
     ResponseEntity<ServiceUsage> getUsage(String serviceInstanceId)
+    ResponseEntity<Set<ServiceUsage>> getExtendedUsage(String serviceInstanceId)
+    ResponseEntity<ServiceHealth> getHealth(String serviceInstanceId)
     ResponseEntity<Void> createOrUpdateServiceDefinition(String definition)
     ResponseEntity<Void> deleteServiceDefinition(String definition)
     ResponseEntity<BackupDto> createBackup(String serviceInstanceId)
@@ -22,4 +25,5 @@ interface IServiceBrokerClientExtended extends IServiceBrokerClient {
     ResponseEntity<String> getApi(String serviceInstanceId)
     ResponseEntity<String> lockUser(String serviceInstanceId)
     ResponseEntity<String> unlockUser(String serviceInstanceId)
+
 }