+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 
+## [2.13.0] - 2018-06-04
+- Fix issue #110 last operation not allowing querying by operationId (client)
+- Add additional extended usage endpoint to service instance
+- Add custom ServiceHealth Endpoint to service instance
+- Fix LoggingInterceptor thread unsafe operation
+
 ## [2.12.1] - 2018-05-07
 - Fix not found binding and instance status code